backend/requirements.txt
fastapi>=0.104.1
jwt>=1.4.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
geoalchemy2>=0.14.2
psycopg[binary]>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
bcrypt>=4.3.0